Move MetadataStorageActionHandler classes to appropriate modules
